refactor: move interfaces, constants into app folder
