+++ src/app/components/sidebar/sidebar.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
-import { UserRoutesType, usersRoute } from '../../../constants/mock-user-data';
-import { UserResponseType } from '../../../interfaces/user.interfaces';
+import { UserRoutesType, usersRoute } from '../../constants/mock-user-data';
+import { UserResponseType } from '../../interfaces/user.interfaces';
 import { AuthStoreService } from '../../services/store/auth-store.service';
 
 @Component({
