+++ src/app/components/stepper/stepper.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnChanges, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { STEPPER_GLOBAL_OPTIONS } from '@angular/cdk/stepper';
-import { Question } from '../../../interfaces/question-answer';
+import { Question } from '../../interfaces/question-answer';
 
 @Component({
   selector: 'app-stepper',
