+++ src/app/interfaces/audition.ts
