+++ src/app/interfaces/essay.ts
