+++ src/app/models/admin-profile/admin-profile-routing.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AdminProfileComponent } from './admin-profile.component';
-import { UserRole } from '../../../constants/data-constants';
+import { UserRole } from '../../constants/data-constants';
 import { AuthGuard } from '../../guard/auth.guard';
 
 const routes: Routes = [
