+++ src/app/models/coach-profile-editor/coach-profile-editor-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { CoachProfileEditorComponent } from './coach-profile-editor.component';
-import { UserRole } from '../../../constants/data-constants';
+import { UserRole } from '../../constants/data-constants';
 import { AuthGuard } from '../../guard/auth.guard';
 
 const routes: Routes = [
