+++ src/app/models/hr-profile/services/post-http.service.ts
@@ -1,8 +1,8 @@
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { Injectable } from '@angular/core';
-import { ApiAssignTest } from '../../../../interfaces/user.interfaces';
-import { PostAssignTest, RefreshUserDataUrl } from '../../../../constants/route-constant';
+import { ApiAssignTest } from '../../../interfaces/user.interfaces';
+import { PostAssignTest, RefreshUserDataUrl } from '../../../constants/route-constant';
 
 @Injectable({
   providedIn: 'root',
