+++ src/app/models/login/login.component.ts
@@ -3,10 +3,10 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Observable } from 'rxjs';
 import { filter } from 'rxjs/operators';
 import { AuthStoreService } from '../../services/store/auth-store.service';
-import { usersMockDataResponse } from '../../../constants/mock-user-data';
-import { UserResponseType } from '../../../interfaces/user.interfaces';
+import { usersMockDataResponse } from '../../constants/mock-user-data';
+import { UserResponseType } from '../../interfaces/user.interfaces';
 import { ErrorStoreService } from '../../services/store/error-store.service';
-import { ErrorModel, ErrorType } from '../../../interfaces/error';
+import { ErrorModel, ErrorType } from '../../interfaces/error';
 
 @Component({
   selector: 'app-login',
