+++ src/app/models/speaking-test/speaking-test.component.ts
@@ -3,8 +3,8 @@ import { DomSanitizer, SafeUrl } from '@angular/platform-browser';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { ErrorStoreService } from '../../services/store/error-store.service';
-import { Route } from '../../../constants/route-constant';
-import { Test } from '../../../interfaces/test';
+import { Route } from '../../constants/route-constant';
+import { Test } from '../../interfaces/test';
 
 declare let MediaRecorder: any;
