+++ src/app/models/user-profile/user-profile-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { UserProfileComponent } from './user-profile.component';
 import { AuthGuard } from '../../guard/auth.guard';
-import { UserRole } from '../../../constants/data-constants';
+import { UserRole } from '../../constants/data-constants';
 
 const routes: Routes = [
   {
