+++ src/app/models/writing-test/writing-test.component.ts
@@ -2,8 +2,8 @@ import { Component, Input, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
-import { Test } from '../../../interfaces/test';
-import { Route } from '../../../constants/route-constant';
+import { Test } from '../../interfaces/test';
+import { Route } from '../../constants/route-constant';
 
 @Component({
   selector: 'app-writing-test',
