+++ src/app/services/test-http.service.ts
@@ -1,14 +1,18 @@
-import { Level } from 'src/constants/data-constants';
-import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { Test } from '../../interfaces/test';
-
-import { BASE_API_URL } from '../../constants/route-constant';
+import { Level } from '../constants/data-constants';
+import { Test } from '../interfaces/test';
+import { BASE_API_URL } from '../constants/route-constant';
+import { Question } from '../interfaces/question-answer';
 
 interface GetTestsResults {
   results: Test[];
+}
+
+interface GetTestContent {
+  grammarQuestions: Question[];
 }
 
 @Injectable({
